Move Editor and Scene folder
