--- Assets/ImageView/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ab19e42c9a4a6964bb1dda4bb4687b00
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
