--- Assets/ImageView/Editor/ImageViewEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e26700309f97e0444ad9e6b10f328516
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
